requirements.txt
certifi==2020.6.20
chardet==3.0.4
click==7.1.2
dnspython==2.1.0
Flask==1.1.2
Flask-Cors==3.0.9
gunicorn==20.1.0
idna==2.10
itsdangerous==1.1.0
Jinja2==2.11.3
MarkupSafe==1.1.1
mtgsdk==1.3.1
pymongo==3.12.0
python-dotenv==0.14.0
PyYAML==5.4
requests==2.26.0
six==1.15.0
urllib3==1.26.5
Werkzeug==1.0.1
